+++ components/esp_hw_support/port/esp32p4/private_include/pmu_param.h
@@ -357,7 +357,7 @@ typedef struct {
         .lp_pad_hold_all = (sleep_flags & PMU_SLEEP_PD_LP_PERIPH) ? 1 : 0, \
         .dig_pause_wdt = ((sleep_flags) & RTC_SLEEP_USE_RTC_WDT) ? 0 : 1, \
     },                                                                  \
-    .icg_func = { 0, clk_flags }                                        \
+    .icg_func = { 0, (uint32_t)(clk_flags) }                            \
 }
 #else // !CONFIG_ESP32P4_SELECTS_REV_LESS_V3
 #define PMU_SLEEP_DIGITAL_DSLP_CONFIG_DEFAULT(sleep_flags, clk_flags) { \
@@ -373,7 +373,7 @@ typedef struct {
         .dig_pad_slp_sel = 0,                                           \
         .dig_pause_wdt = ((sleep_flags) & RTC_SLEEP_USE_RTC_WDT) ? 0 : 1, \
     },                                                                  \
-    .icg_func = { 0, clk_flags }                                        \
+    .icg_func = { 0, (uint32_t)(clk_flags) }                            \
 }
 #endif
